feat: Complete Flickr import system with enhanced metadata and GPS coordinate architecture
## Major Features Added

### Flickr Bulk Import System
- Complete Node.js import pipeline with Cloudinary integration
- Automatic spot creation with proximity-based grouping (100m threshold)
- Enhanced metadata extraction: attribution, capture times, sun position data
- Comprehensive error handling and batch processing with rate limiting
- Support for 334 GPS-enabled photos from Flickr collection

### Backend Sync Service Enhancements
- Bidirectional sync between Supabase and Core Data with conflict resolution
- Enhanced metadata mapping: capture times, attribution, licensing, GPS coordinates
- Integer overflow protection for media dimensions (Int32 bounds checking)
- Automatic coordinate validation and repair for existing data with NaN values
- Smart sync down with 5-minute rate limiting and progress tracking

### Core Data Model Extensions
- New attribution fields: attribution_text, original_source, original_photo_id, license_type
- Enhanced GPS coordinate handling with proper EXIF metadata structure
- Sun snapshot integration with sunrise/sunset calculations
- Selector-based safety checks for model compatibility across versions

### UI/UX Improvements
- Attribution display in Photo Information section with safety checks
- Fixed empty state messaging for Camera & Settings tabs
- Enhanced coordinate validation preventing MapKit crashes
- Automatic sync down on app startup for seamless data caching

### Critical Architecture Fix
- Resolved circular GPS coordinate logic: photos now store original coordinates
- Database stores GPS coordinates in media table (gps_latitude, gps_longitude)
- iOS app downloads photo GPS like other metadata (no more spot coordinate inheritance)
- Eliminates "nan, nan" coordinate display and infinite sun time calculations

## Technical Details

### Database Schema Updates
- Added attribution and GPS fields to media table
- Sun snapshots table integration for cached solar calculations
- PostGIS geography support for spatial queries

### Import Pipeline
- Flickr API integration with metadata preservation
- Cloudinary upload with optimized transformations
- Supabase bulk insert with error recovery
- License mapping and attribution text generation

### Sync Service Robustness
- Multiple integer overflow crash fixes with bounds checking
- Core Data model compatibility using responds(to:) selector checks
- Coordinate validation preventing Invalid Region MapKit errors
- Enhanced debug logging for troubleshooting sync issues

## Files Modified
- iOS: AppState, SyncService, SupabaseManager, SpotDetailView, Core Data models
- Database: New supabase_schema_6_attribution.sql with attribution fields
- Scripts: flickr-bulk-import.js with complete import pipeline
- Config: Package dependencies, environment setup, documentation

## Testing & Validation
- Successfully imported 5+ spots with proper GPS coordinates
- Verified attribution display and enhanced metadata sync
- Tested automatic sync down with rate limiting
- Validated coordinate system preventing NaN crashes

+++ Scenic/Extensions/CDSunSnapshot+Extensions.swift
@@ -15,7 +15,7 @@ extension CDSunSnapshot {
             blueHourStartUTC: self.blueHourStartUTC,
             blueHourEndUTC: self.blueHourEndUTC,
             closestEvent: self.closestEventString.flatMap { SunSnapshot.SolarEvent(rawValue: $0) },
-            relativeMinutesToEvent: self.relativeMinutesToEvent == Int32(Int.max) ? nil : Int(self.relativeMinutesToEvent)
+            relativeMinutesToEvent: self.relativeMinutesToEvent == Int32.max ? nil : Int(self.relativeMinutesToEvent)
         )
     }
     
@@ -35,6 +35,6 @@ extension CDSunSnapshot {
         self.blueHourStartUTC = sunSnapshot.blueHourStartUTC
         self.blueHourEndUTC = sunSnapshot.blueHourEndUTC
         self.closestEventString = sunSnapshot.closestEvent?.rawValue
-        self.relativeMinutesToEvent = Int32(sunSnapshot.relativeMinutesToEvent ?? Int.max)
+        self.relativeMinutesToEvent = Int32(sunSnapshot.relativeMinutesToEvent ?? Int(Int32.max))
     }
 }
